# docs/swagger.yaml
definitions:
...
  dto.CreateUserRequest:
    properties:
      is_toured:
        example: false
        type: boolean
        x-nullable: true
    type: object
  dto.ThreeDimentionalResponse:
    properties:
      id:
        type: string
      path:
        type: string
    type: object
  dto.UserResponse:
    properties:
      id:
        type: string
      is_toured:
        type: boolean
      recorded_model_path:
        type: string
      recorded_voice_path:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: AIRship API
  version: "1.0"
paths:
...
  /users:
    post:
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: body
        name: CreateUserRequest
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - User
  /users/{user_id}:
    delete:
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      tags:
      - User
    get:
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - User
    put:
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      - description: update user
        in: body
        name: CreateUserRequest
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      tags:
      - User
...
  /users/three_dimentionals:
    get:
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ThreeDimentionalResponse'
            type: array
      security:
      - ApiKeyAuth: []
      tags:
      - ThreeDimentionalModel
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      - description: 3dmodel file to be uploaded
        in: formData
        name: ThreeDimentionalModel
        required: true
        type: file
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ThreeDimentionalResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - ThreeDimentionalModel
  /users/three_dimentionals/{three_dimentional_id}:
    delete:
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      - description: ThreeDimentional ID
        in: path
        name: three_dimentional_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      tags:
      - ThreeDimentionalModel
    get:
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      - description: ThreeDimentional ID
        in: path
        name: three_dimentional_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ThreeDimentionalResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - ThreeDimentionalModel
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Bearer [Firebase JWT Token]
        in: header
        name: Authorization
        required: true
        type: string
      - description: ThreeDimentional ID
        in: path
        name: three_dimentional_id
        required: true
        type: string
      - description: 3dmodel file to be uploaded
        in: formData
        name: ThreeDimentionalModel
        required: true
        type: file
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ThreeDimentionalResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - ThreeDimentionalModel
swagger: "2.0"